activeSet(isotone)	R Documentation

Active Set Methods for Isotone Optimization

Description

Isotone optimization can be formulated as a convex programming problem with simple linear constraints. This functions offers active set strategies for a collection of isotone optimization problems pre-specified in the package.

Usage


activeSet(z, isomat, mySolver = lsSolver, ups = 1e-12, check = FALSE, ...)

Arguments

z 	 Vector containing predictors

isomat 	 Matrix with 2 columns that contains isotonicity conditions (see examples)

mySolver 	 Various functions are pre-defined (see details). For user-specified functions fSolver with additional arguments can be used (see details as well).

ups 	 Upper boundary

check 	 If TRUE, KKT feasibility checks for isotonicity of the solution are performed

... 	 Additional arguments for fSolver (see details)

Details

The following solvers are specified: aSolver() for asymmetric least squares, dSolver() for the least absolute value, eSolver() minimizes l1-approximation, fSolver() for arbitrary differentiable functions (additional arguments fobj and gobj must be provided), hSolver() for Huber loss function, iSolver() for SILF loss (support vector regression), lfSolver() for least squares with non-diagonal weights, lsSolver() for least squares with diagonal weights (with response values z and weights w as additional arguments) , mSolver() for Chebyshev norm, oSolver() for power norm, pSolver() for quantile loss function, and finally sSolver() for Poisson likelihood.

See corresponding help files and package vignette for additional description.

Value

Generates an object of class activeset.

x 	 Vector containing the fitted values

lambda 	 Vector with Lagrange multipliers

func.vals 	 Value of the target function

constr.vals 	 Vector with the values of isotonicity constraints

Alambda 	 Constraint matrix multiplied by lambda (should be equal to gradient)

gradient 	 Gradient

isocheck 	 List containing the KKT checks for stationarity, primal feasibility, dual feasibility, and complementary slackness

call 	 Matched call

Author(s)

Jan de Leeuw, Kurt Hornik, Patrick Mair

References

de Leeuw, J., Hornik, K., Mair, P. (2008). Isotone optimization in R: Active Set methods and pool-adjacent-violators algorithm. Journal of Statistical Software, forthcoming.

See Also

gpava, lsSolver, dSolver, mSolver, fSolver, pSolver, lfSolver, oSolver, aSolver, eSolver, sSolver, hSolver, iSolver

Examples



##Example 1: Comparison with PAVA
set.seed(12345)
z <- 9:1                   ##predictor values
y <- rnorm(9)               ##response values
w <- rep(1,9)              ##unit weights
btota <- cbind(1:8, 2:9)    ##Matrix defining isotonicity (total order)

##Now we apply PAVA, LS active set, and, again LS active set by specifying function (fobj) and derivative (gobj)          
fit.pava <- gpava(z, y)
#fit.ls <- activeSet(z, btota, lsSolver, weights = w, y = y)
fit.f <- activeSet(z, btota, fSolver, fobj = function(x) sum(w*(x-y)^2), gobj = function(x) 2*drop(w*(x-y)))

##Matrix with fitted values (basically same results for each method)
#val.mat <- cbind(fit.pava$yfit, fit.ls$x, fit.f$x)
#val.mat

##Example 2: Diagonal vs. non-diagonal weights
set.seed(12345)
ww <- 1:9                                  ##weight vector
wf <- crossprod(matrix(rnorm(81),9,9))/9   ##symmetric weight matrix 

#activeSet(z, btota, lsSolver, check = TRUE, weights = ww, y = y)

#FIXME!!! activeSet(z, btota, lfSolver, check = TRUE, z, weights = wf, y = y) 

##Example 3: Other norms
##Approximate l1 with additional argument eps = 1e-4
activeSet(z, btota, eSolver, check = TRUE, y = y, weights = w, eps = 1e-4)

##Approximate l1 with additional power argument p = 1.2
activeSet(z, btota, oSolver, check = TRUE, y = y, weights = w, p = 1.2)

##Least absolute values (exact l1)
#activeSet(z, btota, dSolver, check = TRUE, weights = w, y = y)

##Approximate l(infty) with additional power argument p = 7
#FIXME!!! activeSet(z, btota, oSolver, check = TRUE, y = y, weights = wu, p = 7)

##Exact l(infty) (Chebyshew norm)
#activeSet(z, btota, mSolver, check = TRUE, weights = w, y = y)

##Poisson likelihood
set.seed(12345)    
y <- rpois(9, 5)
#activeSet(z, btota, sSolver, check = TRUE, weights = w, y = y)

##Asymmetric least squares with additional asymmetric weights aw and bw
activeSet(z, btota, aSolver, check = TRUE, y=y, weights = w, aw = 2, bw = 1)

##Huber norm with additional argument eps = 1
activeSet(z, btota , hSolver, check = TRUE, y = y, weights = w, eps = 1)

##SILF norm with additional arguments beta = 0.8 and eps = 0.2
activeSet(z, btota, iSolver, check = TRUE, y = y, weights = w, beta = 0.8, eps = 0.2)

##Example 4: Other orders
## Tree order
btree <- matrix(c(1,1,2,2,2,3,3,8,2,3,4,5,6,7,8,9),8,2)
#activeSet(z, btree, lsSolver, check = TRUE, weights = w, y = y)

## Block order 
#bprim <- cbind(c(rep(1,3),rep(2,3),rep(3,3),rep(4,3),rep(5,3),rep(6,3)),
#             c(rep(c(4,5,6),3),rep(c(7,8,9),3)))
#activeSet(z, bprim, lsSolver, check = TRUE, weights = wu, y = y)

## Loop order
#bloop <- matrix(c(1,2,3,3,4,5,6,6,7,8,3,3,4,5,6,6,7,8,9,9),10,2)
#activeSet(z, bloop, lsSolver, check = TRUE, weights = wu, y = y)


[Package isotone version 0.3-0 Index]