anacor-package(anacor)	R Documentation

Simple and Canonical Correspondence Analysis

Description

This package performs simple and canonical CA on a two-way frequency table (with missings) by means of SVD. Different scaling methods (standard, centroid, Benzecri, Goodman) as well as various plots including confidence ellipsoids are provided.

Details

Package: 	 anacor

Type: 	 Package

Version: 	 0.9-8

Date: 	 2009-03-03

License: 	 GPL

The main function of the package is anacor() which performs simple CA. The input dataset tab is a (incomplete) frequency table. Various scaling options in terms of standard, centroid, Benzecri, and Goodman scaling are provided. They can be specified using the scaling argument. If covariates for row/column scores are provided, canonical CA will be computed. Utility functions to create various types of indicator matrices are implemented.

Author(s)

Jan de Leeuw, Patrick Mair

Maintainer: Jan de Leeuw <deleeuw@stat.ucla.edu>

References

de Leeuw, J. and Mair, P. (2009). Simple and Canonical Correspondence Analysis Using the R Package anacor. Journal of Statistical Software, forthcoming.

See Also

anacor, <pkg>, <pkg>, <pkg>

Examples


data(tocher)
res <- anacor(tocher)
res
summary(res)

[Package anacor version 0.9-8 Index]