bitterling(corals)	R Documentation

Bitterling

Description

This dataset concerns reproductive behavior of male bitterlings with data derived from 13 sequences using a moving time-window of size two.

Usage

data(bitterling)

Format

A frequency tables with bitterling reproductive behavior at time point 1 (rows) and at time point 2 (columns).

jk

jerking

tu

turning beats

hb

head butting

chs

chasing

ft

fleeing

qu

quivering

le

leading

hdp

head down posture

sk

skimming

sn

snapping

chf

chafing

ffl

finflickering

References

Wiepkema, P.R. (1961). An ethological analysis of the reproductive behavior of the bitterling (rhodeus amarus bloch). Archives Neerlandais Zoologique, 14, 103-199.

Examples


data(bitterling)

[Package corals version 0.2-1 Index]