ekman(smacof)	R Documentation

Ekman data set

Description

Ekman dissimilarities

Usage

data(ekman)

Format

Object of class dist

Details

Ekman presents similarities for 14 colors which are based on a rating by 31 subjects where each pair of colors was rated on a 5-point scale (0 = no similarity up to 4 = identical). After averaging, the similarities were divided by 4 such that they are within the unit interval. Similarities of colors with wavelengths from 434 to 674 nm.

References

Ekman, G. (1954). Dimensions of color vision. Journal of Psychology, 38, 467-474.

Examples


data(ekman)
## maybe str(ekman) ; plot(ekman) ...

[Package smacof version 0.9-2 Index]