glass(corals)	R Documentation

Glass data

Description

Table with occupational status of fathers versus occupational status of their sons for a sample of 3497 British families.

Usage

data(glass)

Format

Rows represent occupation of fathers, columns occupation of sons.

PROF

professional and high administrative

EXEC

managerial and executive

HSUP

higher supervisory

LSUP

lower supervisory

SKIL

skilled manual and routine nonmanual

SEMI

semi-skilled manual

UNSK

unskilled manual

References

Glass, D.V. (1954). Social Mobility in Britain. Glencoe: Free Press.

Examples


data(glass)
## maybe str(glass) ; plot(glass) ...

[Package corals version 0.2-1 Index]