gubell(homals)	R Documentation

Guttman-Bell dataset

Description

This dataset deals with attitudes of social groups.

Usage

gubell

Format

The data frame consists of 7 objects and 5 variables with a total of 17 categories.

References

Guttman, L. (1968), A General Nonmetric Technique for Fitting the Smallest Coordinate Space for a Configuration of Points, Psychometrika, 33, 469-506.

Examples


   data(gubell)
   gubell

[Package homals version 0.8-4 Index]