hartigan(homals)	R Documentation

Hartigan's Hardware

Description

A number of bolts, nails, screws, and tacks are classified according to a number of criteria.

Usage

hartigan

Format

A data frame with the following variables and categories:

Thread: Y = Yes, N = No

Head: F = Flat, C = Cup, O = Cone, R = Round, Y = Cylinder

Head indentation: N = None, T = Star, L = Slit

Bottom: S = Sharp, F = Flat

Length: (in half inches)

Brass: Y = Yes, N = No

References

Hartigan, J.A. (1975). Clustering Algorithms. New York: Wiley.

Examples


   data(hartigan)
   hartigan

[Package homals version 0.8-9 Index]