homals-package(homals)	R Documentation

Homogeneity Analysis

Description

This package performs a homogeneity analysis and various extensions. Rank restrictions on the category quantifications can be imposed (nonlinear PCA). The categories are transformed by means of optimal scaling with options for nominal, ordinal, and numerical scale levels (for rank-1 restrictions). Variables can be grouped into sets, in order to emulate regression analysis and canonical correlation analysis.

Details

Package: 	 homals

Type: 	 Package

Version: 	 0.9-10

Date: 	 2007-02-22

License: 	 GPL

URL: 	 http://r-forge.r-project.org/projects/psychor/

The basic function of the homals package is homals(data) which performs homogeneity analysis and returns an object of class "homals". It allows for grouping variables into sets, constraining ranks on category quantifications and levels. Various plot functions are provided in plot.homals and plot3d.homals.

Author(s)

Jan de Leeuw, Patrick Mair

Maintainer: Jan de Leeuw <deleeuw@ucla.edu>

References

de Leeuw, J., & Mair, P. (2007). Homogeneity analysis in R: The package homals. Preprint available at http://gifi.stat.ucla.edu/homalsR.pdf.

Gifi, A. (1990). Nonlinear Multivariate Analysis. New York: Wiley.

See Also

homals

[Package homals version 0.9-10 Index]