kinshipdelta(smacof)	R Documentation

Kinship Terms

Description

Percentages of how often 15 kinship terms were not grouped together by college students including three external scales.

Usage

data(kinshipdelta)

data(kinshipscales)

Format

Dissimilarity matrix of 15 kinship terms and data frame with the following external scales:

Gender (1 = male, 2 = female, 9 = missing)

Generation (-2 = two back, -1 = one back, 0 = same generation, 1 = one ahead, 2 = two ahead)

Degree (1 = first, 2 = second, 3 = third, 4 = fourth)

References

Rosenberg, S. & Kim, M. P. (1975). The method of sorting as a data gathering procedure in multivariate research. Multivariate Behavioral Research, 10, 489-502.

Examples



data(kinshipdelta)
data(kinshipscales)


[Package smacof version 0.0-9 Index]