mkIndiList(anacor)	R Documentation

Converts Table to Indicator Matrix

Description

This function takes a data frame, a vector of types, a list of knot vectors, and a vector of orders. It returns a list of codings for the variables, i.e., crisp indicator, numerical version, or fuzzy indicator.

Usage


mkIndiList(data, type = rep("C",dim(data)[2]), knots = repList(0,length(type)),
ord = rep(1,length(type)))

Arguments

data 	 Data frame to be converted.

type 	 If "C", a crisp indicator is returned; if "A", a numerical version; if "F", the b-spline basis as a fuzzy indicator.

knots 	 Knot sequence for type-C coding.

ord 	 Variable order for type-C coding.

See Also

expandFrame, burtTable

Examples



# artificial data 
x1 <- sample(1:6, 20, replace = TRUE) 
x2 <- sample(1:3, 20, replace = TRUE)
data <- data.frame(x1,x2)
mkIndiList(data)


[Package anacor version 0.9-8 Index]