neumann(homals)	R Documentation

Neumann dataset

Description

Willard Gibbs discovered a theoretical formula connecting the density, the pressure, and the absolute temperature of a mixture of gases with convertible components. He applied this formula and the estimated constants to 65 experiments of Neumann, and he discusses the systematic and accidental divergences (residuals).

Usage

neumann

Format

A data frame with temperature, pressure, and density on 65 experiments.

References

Gifi, A. (1990). Nonlinear multivariate analysis. New York: Wiley. Wilson, E.B. (1926). Empiricism and rationalism. Science, 64, 47-57

Examples


   data(neumann)
   neumann

[Package homals version 0.8-5 Index]