plot.homals(homals)	R Documentation

Plot methods for homals

Description

These methods provide various 2D and 3D plots for objects of class "homals".

Usage


## S3 method for class 'homals':
plot(x, plot.dim = c(1, 2), plot.type, var.subset, main, type, xlab, ylab, xlim, ylim, 
leg.pos = "topright", ...)

## S3 method for class 'homals':
plot3d(x, plot.dim = c(1, 2, 3), plot.type, var.subset, type, xlab, ylab, 
zlab, col, main, sphere = TRUE, bgpng = "particle.png", ax.grid = TRUE, ...)

## S3 method for class 'homals':
plot3dstatic(x, plot.dim = c(1, 2, 3), plot.type, var.subset, main, type, 
xlab, ylab, zlab, xlim, ylim, zlim, ...)


Arguments

x 	 Object of class "homals"

plot.dim 	 Vector with dimensions to be plotted against.

plot.type 	 String indicating which type of plot to be produced: "jointplot", "catplot" "labplot", "objplot", "starplot", "graphplot", "hullplot", "lossplot", "prjplot", "spanplot", , "trfplot", "vecplot", "vorplot", "loadplot" (Not all of these are provided as 3d-plots, see below).

var.subset 	 Numeric vector for subsetting variables to be plotted. If missing, all variables are taken into account. Ignored for non separate variable plots.

main 	 Plot title.

type 	 Type of points/lines to be plotted.

xlab 	 Label of x-axis.

ylab 	 Label of y-axis.

zlab 	 Label of z-axis.

col 	 Plotting color.

xlim 	 Limits for x-axis.

ylim 	 Limits for y-axis.

zlim 	 Limits for z-axis.

leg.pos 	 Position of the legend (for "trfplot" and "spanplot" only) as provided in legend.

sphere 	 Whether 3D-sphere should be drawn.

bgpng 	 Background image in /library/rgl/textures/. If NULL, blank background.

ax.grid 	 If TRUE grid is drawn.

... 	 Further plot arguments passed: see plot, plot3d in package rgl, and scatterplot3d in package scatterplot3d for detailed information.

Details

Plot description:

- Object plot (plot.type = "objplot"): Plots the scores of the objects (rows in data set) on two or three dimensions.

- Category plot (plot.type = "catplot"): Plots the rank-restricted category quantifications for each variable separately. Three-dimensional plot is available.

- Voronoi plot (plot.type = "vorplot"): Produces a category plot with Voronoi regions.

- Joint plot (plot.type = "jointplot"): Plots the object scores and the category quantifications in 1 device. Three-dimensional version is available.

- Graph plot (plot.type = "graphplot"): Joint plot with connections between scores/quantifications.

- Hull plot (plot.type = "hullplot"): For each single variable the object scores are mapped onto two dimensions and the convex hull for each response category is drawn.

- Label plot (plot.type = "labplot"): Similar to object plot, the object scores are plotted but for each variable separately with the corresponding category labels. A three-dimensional version is provided.

- Span plot (plot.type = "spanplot"): As label plot, it maps the object scores for each variable and span plot connects them by the shortest path within each response category.

- Star plot (plot.type = "starplot"): Again, the object scores are mapped. In addition these points are connected with the category centroid. Three-dimensional version is provided.

- Loss plot (plot.type = "lossplot"): Plots the rank-restricted category quantifications against the unrestricted for each variable separately.

- Projection plot (plot.type = "prjplot"): For variables of rank 1 the object scores (two-dimensional) are projected onto a straight line determined by the rank restricted category quantifications.

- Vector plot (plot.type = "vecplot"): For variable of rank 1 the object scores (two-dimensional) are projected onto a straight line determined by the rank restricted category quantifications.

- Transformation plot (plot.type = "trfplot"): Plots the original (categorical) scale against the transformed (metric) scale on each dimension over the categories of each variable separately.

- Loadings plot (plot.type = "loadplot"): Plots the loadings of the variables and connects them with the origin. Three-dimensional version is available.

See Also

homals

Examples



##Graphplot for Hartigan solution: 3 dimensions extracted, dimension 1 plotted
##against dimension 3.
data(hartigan)
res <- homals(hartigan, ndim = 3)
plot(res, plot.dim = c(1,3), plot.type = "graphplot")

##3D star plot for indentation and bottom
plot3d(res, plot.type = "starplot", var.subset = c(2,4))

##Static 3D joint plot for thread and bottom
plot3dstatic(res, plot.type = "jointplot", var.subset = c(1,4))

[Package homals version 0.9-3 Index]