plot.smacof(smacof)	R Documentation

2D SMACOF plots

Description

These methods provide various 2D plots for SMACOF models.

Usage


## S3 method for class 'smacof':
plot(x, plot.type = "confplot", plot.dim = c(1,2), sphere = TRUE, 
main, xlab, ylab, xlim, ylim, ...)

## S3 method for class 'smacofR':
plot(x, plot.type = "confplot", joint = FALSE, plot.dim = c(1,2), 
main, xlab, ylab, xlim, ylim, ...)

## S3 method for class 'smacofID':
plot(x, plot.type = "confplot", plot.dim = c(1,2), main, xlab, ylab, xlim, ylim, ...)


Arguments

x 	 Object of class "smacof", "smacofR", and "smacofID" (see details)

plot.type 	 String indicating which type of plot to be produced: "confplot", "resplot" "Shepard", "stressplot" (see details)

plot.dim 	 Vector with dimensions to be plotted.

main 	 Plot title.

xlab 	 Label of x-axis.

ylab 	 Label of y-axis.

xlim 	 Scale x-axis.

ylim 	 Scale y-axis.

sphere 	 In case of spherical smacof, whether sphere should be plotted or not.

joint 	 If TRUE, the configurations are plotted jointly in rectangular smacof.

... 	 Further plot arguments passed: see plot

in package scatterplot3d for detailed information.

Details

smacofSym() creates object of class "smacof", whereas smacofRect() produces "smacofR" and smacofIndDiff() generates "smacofID".

Plot description:

- Configuration plot (plot.type = "confplot"): Plots the MDS configurations.

- Residual plot (plot.type = "resplot"): Plots the configuration distances against the corresponding residuals.

- Shepard diagram (plot.type = "Shepard"): Diagram with the observed against the fitted distances including isotonic regression line.

- Stress decomposition plot (plot.type = "stressplot"): Plots the stress contribution in of each observation.

For smacofIndDiff() the residual plot, Shepard diagram, and stress plot are based on the sum of the residuals across individuals/ways. The configuration plot represents the group stimulus space (i.e., joint configurations).

See Also

plot3d.smacof

Examples



## 2D plots for spherical SMACOF
data(trading)
res <- smacofSphere.dual(trading)
plot(res, plot.type = "confplot", sphere = TRUE)
plot(res, plot.type = "resplot")
plot(res, plot.type = "Shepard")
plot(res, plot.type = "stressplot")

## Joint configuration plot and row/column stressplots for rectangular SMACOF
data(breakfast)
res <- smacofRect(breakfast)
plot(res, plot.type = "confplot", joint = TRUE)
plot(res, plot.type = "stressplot")


[Package smacof version 0.9-3 Index]