predict.homals(homals)	R Documentation

Classification tables for homals solutions

Description

Given a multiple homals solution we can reconstruct the indicator matrix by assigning each object to the closest category point of the variable. We can then find out how well we have reconstructed the original data in terms of a classification table.

Usage


## S3 method for class 'homals':
predict(object, ...)

Arguments

object 	 Object of class "homals"

... 	 Further arguments ignored

Details

For variables with rank restrictions we first project the objects on the hyperplane spanned by the category quantifications, and then compute distances in that plane. If there are sets of variables, the percentage correctly classified for a variable is usually not the best thing to look at; except in some special cases, in which some sets consist of a single variable (such as regression and discriminant analysis). In any case a square table with observed vs. predicted values is computed.

Value

cl.table 	 List of classification tables for each variable.

cr.vec 	 Vector with classification rates.

See Also

homals

Examples


  data(galo)
  res <- homals(galo, active=c(TRUE, TRUE,TRUE, TRUE, FALSE))
  pr.res <- predict(res)
  pr.res
  summary(pr.res)

[Package homals version 0.8-10 Index]