print.homals(homals)	R Documentation

S3 methods for homals objects

Description

Print and summary method for objects of class "homals"

Usage


## S3 method for class 'homals':
print(x, ...)
## S3 method for class 'homals':
summary(object, ...)

Arguments

x 	 Object of class "homals".

object 	 Object of class "homals".

... 	 Further arguments are ignored

Details

The print methods provides the eigenvalues and the rank-restricted category quantifications for each variable. The summary method gives the caterogy loadings, category centroids, rank-restricted category quantifications, and lower rank quantifications for each variable separately.

Examples



data(hartigan)
res <- homals(hartigan)
res
summary(res)


[Package homals version 0.8-10 Index]