sim2diss(smacof)	R Documentation

Converts similarites to dissimilarities

Description

Utility function for converting similarities into dissimilarities. Different methods are provided.

Usage


sim2diss(similmat, method = "corr", to.dist = TRUE)

Arguments

similmat 	 Similarity matrix (not necessarily symmetric, nor quadratic)

method 	 Various methods for converting similarities into dissimilarities: "corr", "neglog", "counts", or an integer value (see details)

to.dist 	 If TRUE, object of class dist is produced

Details

We provide the following methods for converting similarities S into dissimilarities D: "corr" is suited for correlation matrices and takes D = sqrt(1-S). "neglog" takes the negative logarithm in terms of -log(S). Having frequencies, "counts" is appropriate which does -log((S[i,j]*S[j,i)/(S[i,i]*S[j,j])). The user can specify also an integer value v. In this case sim2diss() computes v-S.

Value

Returns dissimiarities either as matrix or as dist object.

Examples



## Converting Ekman data (similarities) into dissimilarities by subtraction from 1
data(ekman)
ekman.diss <- sim2diss(ekman, method = 1)
res <- smacofSym(ekman.diss)

[Package smacof version 0.0-9 Index]