smacof-package(smacof)	R Documentation

SMACOF for Multidimensional Scaling.

Description

This package provides the following approaches of multidimensional scaling (MDS) based on stress minimization by means of majorization (smacof): Simple smacof on symmetric dissimilarity matrices, smacof for rectangular matrices (unfolding models), smacof with constraints on the configuration, three-way smacof for individual differences (including constraints for idioscal, indscal, and identity), and spherical smacof (primal and dual algorithm). Each of these approaches is implemented in a metric and nonmetric manner including primary, secondary, and tertiary approaches for tie handling.

Details

Package: 	 smacof

Type: 	 Package

Version: 	 0.9-6

Date: 	 2008-05-13

License: 	 GPL

The function for basic SMACOF on symmetric dissimilarity matrices is smacofSym(). For rectangular input matrices (unfolding model) smacofRect() is appropriate and by means of smacofIndDiff() individual difference models (three-way MDS) can be computed.

Author(s)

Jan de Leeuw, Patrick Mair

Maintainer: Jan de Leeuw <deleeuw@stat.ucla.edu>

References

de Leeuw, J. & Mair, P. (2008). Multidimensional scaling using majorization: The R package smacof. Journal of Statistical Software, forthcoming.

See Also

smacofSym, smacofRect, smacof, smacofIndDiff, smacofSphere.primal, smacofSphere.dual

Examples


data(trading)
res <- smacofSym(trading)
res

[Package smacof version 0.9-6 Index]