smacofConstraint(smacof)	R Documentation

SMACOF Constraint

Description

SMACOF with constraints on the configuration

Usage


smacofConstraint(delta, constraint = "linear", external, ndim = 2, weightmat = NULL, 
metric = TRUE, ties = "primary", verbose = FALSE, modulus = 1, itmax = 1000, eps = 1e-6)


Arguments

delta 	 Either a symmetric dissimilarity matrix or an object of class "dist"

constraint 	 Type of constraint: "linear", "unique", "diagonal", or a user-specified function (see details)

external 	 Data frame or matrix with external covariates

ndim 	 Number of dimensions

weightmat 	 Optional matrix with dissimilarity weights

metric 	 If FALSE non-metric MDS is performed

ties 	 Tie specification for non-metric MDS only: "primary", "secondary", or "tertiary"

verbose 	 If TRUE, intermediate stress is printed out

modulus 	 Number of smacof iterations per monotone regression call

itmax 	 Maximum number of iterations

eps 	 Convergence criterion

Details

The user can specify a function with the following arguments: configuration matrix (n x p), matrix V (based on the weight matrix, see vignette), external scale matrix. The function must return a (n x p) matrix of resulting configurations. In the examples we show how linearly restricted configurations can be specified by the user.

Value

obsdiss 	 Observed dissimilarities, normalized

confdiss 	 Configuration dissimilarities

conf 	 Matrix of final configurations

stress.m 	 stress value for metric MDS

stress.nm 	 stress value for non-metric MDS (if computed)

ndim 	 Number of dimensions

model 	 Type of smacof model

niter 	 Number of iterations

nobj 	 Number of objects

Author(s)

Jan de Leeuw and Patrick Mair

References

de Leeuw, J. & Mair, P. (2008). Multidimensional scaling using majorization: The R package smacof.

See Also

smacofSym, smacofRect, smacofIndDiff, smacofSphere.primal, smacofSphere.dual

Examples



## SMACOF with linear configuration constraints
data(kinshipdelta)
data(kinshipscales)
res.lin1 <- smacofConstraint(kinshipdelta, constraint = "linear", external = kinshipscales)

## The same model, linear restrictions specified externally
## constrfun <- function(X, V, external) {
##      return(external
## }
## res.lin2 <- smacofConstraint(kinshipdelta, constraint = constrfun, external = kinshipscales)


[Package smacof version 0.0-10 Index]