smacofIndDiff(smacof)	R Documentation

smacof for Individual Differences

Description

Performs smacof for individual differences also known as Three-Way smacof on a list of dissimilarity matrices. Various restrictions decompositions and restrictions on the weight matrix are provided.

Usage


smacofIndDiff(delta, ndim = 2, weightmat = NULL, init = NULL, metric = TRUE,
              ties = "primary", constraint = NULL, verbose = FALSE, modulus = 1,
              itmax = 1000, eps = 1e-6)


Arguments

delta 	 A list of dissimilarity matrices or a list objects of class dist

ndim 	 Number of dimensions

weightmat 	 Optional matrix with dissimilarity weights

init 	 Matrix with starting values for configurations (optional)

metric 	 If FALSE non-metric MDS is performed

ties 	 Tie specification for non-metric MDS

constraint 	 Either NULL, "idioscal", "diagonal", or "identity" (see details)

verbose 	 If TRUE, intermediate stress is printed out

modulus 	 Number of smacof iterations per monotone regression call

itmax 	 Maximum number of iterations

eps 	 Convergence criterion

Details

If the constraint is NULL, INDSCAL is performed with identity configuration weight matrices. An additional restriction can be imposed with "identity" which restricts the configurations across individuals/replications/ways to be equal. More unrestricted models are "diagonal" which restricts only the configuration weight matrices to be diagonal and "idioscal" is unrestricted.

Value

obsdiss 	 List of observed dissimilarities, normalized

confdiss 	 List of configuration dissimilarities

conf 	 List of matrices of final configurations

gspace 	 Joint configurations aka group stimulus space

cweights 	 Configuration weights

stress.m 	 stress value for metric MDS

stress.nm 	 stress value for non-metric MDS (if computed)

stress.co 	 Constrained stress value

ndim 	 Number of dimensions

model 	 Type of smacof model

niter 	 Number of iterations

nobj 	 Number of objects

Author(s)

Jan de Leeuw and Patrick Mair

References

de Leeuw, J. & Mair, P. (2008). Multidimensional scaling using majorization: The R package smacof.

See Also

smacofConstraint, smacofSym, smacofRect, smacofSphere.primal, smacofSphere.dual

Examples


data(perception)
res <- smacofIndDiff(perception)
res
summary(res)

res.id <- smacofIndDiff(perception, constraint = "identity")
res.diag <- smacofIndDiff(perception, constraint = "diagonal")
res.idio <- smacofIndDiff(perception, constraint = "idioscal")

[Package smacof version 0.9-4 Index]