spider(corals)	R Documentation

Hunting spider data

Description

Abundance of hunting spiders in a Dutch dune area.

Usage

data(glass)

Format

A list of data frames containing the frequency table (28 observations) and the row covariates.

Table:

Alopacce

Abundance of Alopecosa accentuata.

Alopcune

Abundance of Alopecosa cuneata.

Alopfabr

Abundance of Alopecosa fabrilis.

Arctlute

Abundance of Arctosa lutetiana.

Arctperi

Abundance of Arctosa perita.

Auloalbi

Abundance of Aulonia albimana.

Pardlugu

Abundance of Pardosa lugubris.

Pardmont

Abundance of Pardosa monticola.

Pardnigr

Abundance of Pardosa nigriceps.

Pardpull

Abundance of Pardosa pullata.

Trocterr

Abundance of Trochosa terricola.

Zoraspin

Abundance of Zora spinimana.

Row covariates:

WaterCon

Log percentage of soil dry mass.

BareSand

Log percentage cover of bare sand.

FallTwig

Log percentage cover of fallen leaves and twigs.

CoveMoss

Log percentage cover of the moss layer.

CoveHerb

Log percentage cover of the herb layer.

ReflLux

Reflection of the soil surface with cloudless sky.

References

Van der Aart, P.J.M. and Smeek-Enserink, N. (1975). Correlations between distributions of hunting spiders (Lycosidae, Ctenidae) and environmental characteristics in a dune area. Netherlands Journal of Zoology, 25, 1–45.

Examples


data(spider)
str(spider)

[Package corals version 0.2-1 Index]