srole(anacor)	R Documentation

Srole Data

Description

This dataset provides a cross-classification of subjects according to their mental health status and parents' socio-economic status.

Usage

data(srole)

Format

Mental health has four categories (rows): well, mild symptom formation, moderate symptom formation, and impaired. There are six categories of socio-economic status in the columns.

References

Srole, L., Langner, T.S., Michael, S.T., Opler, M.K., & Rennie, T.A.C. (1962). Mental health in the metropolis: The midtown Manhattan study. New York: McGraw-Hill.

Examples


data(srole)
## maybe str(srole) ; plot(srole) ...

[Package anacor version 0.4-0 Index]