homals-package(homals)	R Documentation

Homogeneity Analysis

Description

This package performs a homogeneity analysis, aka a multiple correspondence analysis, but with many additional options. Variables can be grouped into sets, in order to emulate regression analysis and canonical analysis. For each variable there are, in addition, rank constraints on the category quantifications (or transformations) and level constraints (which allows one to treat a variable as nominal, ordinal, or numerical).

Details

Package: 	 homals

Type: 	 Package

Version: 	 0.8-10

Date: 	 2007-10-15

License: 	 GPL

The basic function of the homals package is homals(data) which performs homogeneity analysis and returns an object of class "homals". It allows for grouping variables into sets, constraining ranks on category quantifications and levels. Various plot functions are provided in plot.homals and plot3d.homals.

Author(s)

Jan de Leeuw, Patrick Mair

Maintainer: Jan de Leeuw <deleeuw@ucla.edu>

References

Gifi, A. (1990). Nonlinear Multivariate Analysis. New York: Wiley.

See Also

homals

[Package homals version 0.8-10 Index]